+++ idf_build_apps/main.py
@@ -25,6 +25,7 @@ from .manifest.manifest import (
 )
 from .utils import (
     BuildError,
+    files_matches_patterns,
     get_parallel_start_stop,
     to_list,
 )
@@ -52,6 +53,8 @@ def find_apps(
     default_build_targets=None,  # type: list[str] | str | None
     depends_on_components=None,  # type: list[str] | str | None
     manifest_rootpath=None,  # type: str | None
+    ignore_component_dependencies_file_patterns=None,  # type: list[str] | str | None
+    depends_on_files=None,  # type: list[str] | str | None
 ):  # type: (...) -> list[App]
     if default_build_targets:
         default_build_targets = to_list(default_build_targets)
@@ -69,6 +72,23 @@ def find_apps(
         manifest = Manifest(rules)
         App.MANIFEST = manifest
 
+    depends_on_components = to_list(depends_on_components)
+    if depends_on_components is None:
+        check_component_dependencies = False
+    elif (
+        ignore_component_dependencies_file_patterns
+        and depends_on_files
+        and files_matches_patterns(depends_on_files, ignore_component_dependencies_file_patterns)
+    ):
+        LOGGER.debug(
+            'Skipping check component dependencies for apps since files %s matches patterns: %s',
+            ', '.join(depends_on_files),
+            ', '.join(ignore_component_dependencies_file_patterns),
+        )
+        check_component_dependencies = False
+    else:
+        check_component_dependencies = True
+
     apps = []
     if target == 'all':
         targets = ALL_TARGETS
@@ -81,9 +101,9 @@ def find_apps(
                 _find_apps(
                     path,
                     target,
-                    build_system=build_system,
-                    recursive=recursive,
-                    exclude_list=exclude_list or [],
+                    build_system,
+                    recursive,
+                    exclude_list or [],
                     work_dir=work_dir,
                     build_dir=build_dir or 'build',
                     config_rules_str=config_rules_str,
@@ -92,6 +112,7 @@ def find_apps(
                     check_warnings=check_warnings,
                     preserve=preserve,
                     depends_on_components=depends_on_components,
+                    check_component_dependencies=check_component_dependencies,
                 )
             )
     apps.sort()
@@ -113,6 +134,9 @@ def build_apps(
     ignore_warning_file=None,  # type: t.TextIO | None
     copy_sdkconfig=False,  # type: bool
     depends_on_components=None,  # type: list[str] | str | None
+    manifest_rootpath=None,  # type: str | None
+    ignore_component_dependencies_file_patterns=None,  # type: list[str] | str | None
+    depends_on_files=None,  # type: list[str] | str | None
 ):  # type: (...) -> (int, list[App]) | int
     apps = to_list(apps)
 
@@ -124,6 +148,23 @@ def build_apps(
         for s in ignore_warning_file:
             ignore_warnings_regexes.append(re.compile(s.strip()))
     App.IGNORE_WARNS_REGEXES = ignore_warnings_regexes
+
+    depends_on_components = to_list(depends_on_components)
+    if depends_on_components is None:
+        check_component_dependencies = False
+    elif (
+        ignore_component_dependencies_file_patterns
+        and depends_on_files
+        and files_matches_patterns(depends_on_files, ignore_component_dependencies_file_patterns, manifest_rootpath)
+    ):
+        LOGGER.debug(
+            'Skipping check component dependencies for apps since files %s matches patterns: %s',
+            ', '.join(depends_on_files),
+            ', '.join(ignore_component_dependencies_file_patterns),
+        )
+        check_component_dependencies = False
+    else:
+        check_component_dependencies = True
 
     start, stop = get_parallel_start_stop(len(apps), parallel_count, parallel_index)
     LOGGER.info('Total %s apps. running build for app %s-%s', len(apps), start, stop)
@@ -152,7 +193,10 @@ def build_apps(
         LOGGER.info('Building app %s: %s', index, repr(app))
         is_built = False
         try:
-            is_built = app.build(depends_on_components)
+            is_built = app.build(
+                depends_on_components=depends_on_components,
+                check_component_dependencies=check_component_dependencies,
+            )
         except BuildError as e:
             LOGGER.error(str(e))
             if keep_going: